# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Outbreak Trust verification service API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SourceStatusKind = typeof SourceStatusKind[keyof typeof SourceStatusKind];


export const SourceStatusKind = {
  source: 'source',
  chain: 'chain',
} as const;

export type SourceStatusStatus = typeof SourceStatusStatus[keyof typeof SourceStatusStatus];


export const SourceStatusStatus = {
  connected: 'connected',
  pending: 'pending',
  unavailable: 'unavailable',
  demo: 'demo',
} as const;

export interface SourceStatus {
  id: string;
  label: string;
  kind: SourceStatusKind;
  status: SourceStatusStatus;
  detail: string;
  lastCheckedAt: string;
  /** @nullable */
  network: string | null;
  /** @nullable */
  explorerUrl: string | null;
}

export interface VerificationSummary {
  totalRuns: number;
  anchoredRuns: number;
  sourceAgreement: number;
  recordsCompared: number;
  /** @nullable */
  lastVerifiedAt: string | null;
  /** @nullable */
  latestHash: string | null;
  /** @nullable */
  latestTxHash: string | null;
}

export interface VerificationRunInput {
  /** @minLength 2 */
  geography: string;
  /** @minLength 2 */
  indicator: string;
  /** @minLength 4 */
  period: string;
  anchor?: boolean;
}

export type VerificationRunStatus = typeof VerificationRunStatus[keyof typeof VerificationRunStatus];


export const VerificationRunStatus = {
  anchored: 'anchored',
  verified: 'verified',
  'needs-review': 'needs-review',
  failed: 'failed',
} as const;

export type SourceResultStatus = typeof SourceResultStatus[keyof typeof SourceResultStatus];


export const SourceResultStatus = {
  matched: 'matched',
  diverged: 'diverged',
  unavailable: 'unavailable',
  demo: 'demo',
} as const;

export interface SourceResult {
  source: string;
  status: SourceResultStatus;
  /** @nullable */
  value: number | null;
  unit: string;
  observedAt: string;
  note: string;
}

export interface VerificationRun {
  id: number;
  geography: string;
  indicator: string;
  period: string;
  status: VerificationRunStatus;
  agreement: number;
  recordsCompared: number;
  sourceResults: SourceResult[];
  report: string;
  contentHash: string;
  /** @nullable */
  txHash: string | null;
  /** @nullable */
  explorerUrl: string | null;
  createdAt: string;
}

export type ListVerificationRunsParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};