# lib/api-zod/src/generated/types/sourceResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Outbreak Trust verification service API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceResultStatus } from './sourceResultStatus';

export interface SourceResult {
  source: string;
  status: SourceResultStatus;
  /** @nullable */
  value: number | null;
  unit: string;
  observedAt: string;
  note: string;
}