# lib/api-zod/src/generated/types/sourceStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Outbreak Trust verification service API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceStatusKind } from './sourceStatusKind';
import type { SourceStatusStatus } from './sourceStatusStatus';

export interface SourceStatus {
  id: string;
  label: string;
  kind: SourceStatusKind;
  status: SourceStatusStatus;
  detail: string;
  lastCheckedAt: Date;
  /** @nullable */
  network: string | null;
  /** @nullable */
  explorerUrl: string | null;
}