# lib/api-zod/src/generated/types/verificationRun.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Outbreak Trust verification service API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceResult } from './sourceResult';
import type { VerificationRunStatus } from './verificationRunStatus';

export interface VerificationRun {
  id: number;
  geography: string;
  indicator: string;
  period: string;
  status: VerificationRunStatus;
  agreement: number;
  recordsCompared: number;
  sourceResults: SourceResult[];
  report: string;
  contentHash: string;
  /** @nullable */
  txHash: string | null;
  /** @nullable */
  explorerUrl: string | null;
  createdAt: Date;
}