# lib/api-zod/src/generated/types/verificationRunInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Outbreak Trust verification service API
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationRunInput {
  /** @minLength 2 */
  geography: string;
  /** @minLength 2 */
  indicator: string;
  /** @minLength 4 */
  period: string;
  anchor?: boolean;
}