# lib/api-zod/src/generated/types/verificationSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Outbreak Trust verification service API
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationSummary {
  totalRuns: number;
  anchoredRuns: number;
  sourceAgreement: number;
  recordsCompared: number;
  /** @nullable */
  lastVerifiedAt: string | null;
  /** @nullable */
  latestHash: string | null;
  /** @nullable */
  latestTxHash: string | null;
}